refactor: 인터페이스 위치, 내부로 이동
